requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-oauth-toolkit==1.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-social-oauth2==1.2.1
drf-yasg==1.20.0
ecdsa==0.17.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
Markdown==3.3.4
MarkupSafe==2.1.0
oauthlib==3.2.0
packaging==21.3
Pillow==8.4.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-decouple==3.5
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.3
razorpay==1.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shortuuid==1.0.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
